requirements.txt
boto3
pandas
PyMySQL
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
gspread-dataframe